chore(create-email): Move the starter creation into a new file
